Let state monitor manager loop handle deletion in all cases
With the way StateMonitorManager is designed, deletion will leave
the item in the priority queue. Also makes an issue if it is started
before it can be deleted from the priority queue. This is a bit simpler
and safer way to handle deletions, and is the exact same for end-users